Rename AttachmentData#virus_scan_pending?
The `&& true` clause in `virus_scan_pending?` is redundant and the
method can be renamed to reflect what it's doing. I've also changed
the test that exercised this behaviour to reflect the change
realistic.

After this refactor, it looks like this clause is redundant so I will
remove it in a subsequent commit.